+++ packages/mobx-react-lite/__tests__/strictAndConcurrentModeUsingFinalizationRegistry.test.tsx
@@ -24,6 +24,7 @@ async function gc_cycle() {
 }
 
 test("uncommitted components should not leak observations", async () => {
+    jest.setTimeout(30_000)
     const store = mobx.observable({ count1: 0, count2: 0 })
 
     // Track whether counts are observed
@@ -52,7 +53,6 @@ test("uncommitted components should not leak observations", async () => {
 
     // Allow gc to kick in in case to let finalization registry cleanup
     await gc_cycle()
-
     // Can take a while (especially on CI) before gc actually calls the registry
     await waitFor(
         () => {
@@ -62,7 +62,7 @@ test("uncommitted components should not leak observations", async () => {
             expect(count2IsObserved).toBeFalsy()
         },
         {
-            timeout: 3000,
+            timeout: 10_000,
             interval: 200
         }
     )