+++ packages/mobx-react/__tests__/hooks.test.tsx
@@ -93,6 +93,7 @@ test("computed properties result in double render when using observer instead of
     expect(seen).toEqual([
         "parent",
         0,
+        0,
         "parent",
         2,
         2 // should contain "2" only once! But with hooks, one update is scheduled based the fact that props change, the other because the observable source changed.