+++ packages/mobx-react/src/observerClass.ts
@@ -1,12 +1,10 @@
 import { PureComponent, Component, ComponentClass, ClassAttributes } from "react"
 import {
-    createAtom,
     _allowStateChanges,
     Reaction,
     _allowStateReadsStart,
     _allowStateReadsEnd,
-    _getGlobalState,
-    IAtom
+    _getGlobalState
 } from "mobx"
 import {
     isUsingStaticRendering,
@@ -17,21 +15,25 @@ import { shallowEqual, patch } from "./utils/utils"
 const administrationSymbol = Symbol("ObserverAdministration")
 const isMobXReactObserverSymbol = Symbol("isMobXReactObserver")
 
+let observablePropDescriptors: PropertyDescriptorMap
+if (__DEV__) {
+    observablePropDescriptors = {
+        props: createObservablePropDescriptor("props"),
+        state: createObservablePropDescriptor("state"),
+        context: createObservablePropDescriptor("context")
+    }
+}
+
 type ObserverAdministration = {
     reaction: Reaction | null // also serves as disposed flag
     forceUpdate: Function | null
     mounted: boolean // we could use forceUpdate as mounted flag
+    reactionInvalidatedBeforeMount: boolean
     name: string
-    propsAtom: IAtom
-    stateAtom: IAtom
-    contextAtom: IAtom
+    // Used only on __DEV__
     props: any
     state: any
     context: any
-    // Setting this.props causes forceUpdate, because this.props is observable.
-    // forceUpdate sets this.props.
-    // This flag is used to avoid the loop.
-    isUpdating: boolean
 }
 
 function getAdministration(component: Component): ObserverAdministration {
@@ -41,15 +43,12 @@ function getAdministration(component: Component): ObserverAdministration {
     return (component[administrationSymbol] ??= {
         reaction: null,
         mounted: false,
+        reactionInvalidatedBeforeMount: false,
         forceUpdate: null,
         name: getDisplayName(component.constructor as ComponentClass),
         state: undefined,
         props: undefined,
-        context: undefined,
-        propsAtom: createAtom("props"),
-        stateAtom: createAtom("state"),
-        contextAtom: createAtom("context"),
-        isUpdating: false
+        context: undefined
     })
 }
 
@@ -60,9 +59,8 @@ export function makeClassComponentObserver(
 
     if (componentClass[isMobXReactObserverSymbol]) {
         const displayName = getDisplayName(componentClass)
-        console.warn(
-            `The provided component class (${displayName})
-                has already been declared as an observer component.`
+        throw new Error(
+            `The provided component class (${displayName}) has already been declared as an observer component.`
         )
     } else {
         componentClass[isMobXReactObserverSymbol] = true
@@ -82,15 +80,9 @@ export function makeClassComponentObserver(
         }
     }
 
-    // this.props and this.state are made observable, just to make sure @computed fields that
-    // are defined inside the component, and which rely on state or props, re-compute if state or props change
-    // (otherwise the computed wouldn't update and become stale on props change, since props are not observable)
-    // However, this solution is not without it's own problems: https://github.com/mobxjs/mobx-react/issues?utf8=%E2%9C%93&q=is%3Aissue+label%3Aobservable-props-or-not+
-    Object.defineProperties(prototype, {
-        props: observablePropsDescriptor,
-        state: observableStateDescriptor,
-        context: observableContextDescriptor
-    })
+    if (__DEV__) {
+        Object.defineProperties(prototype, observablePropDescriptors)
+    }
 
     const originalRender = prototype.render
     if (typeof originalRender !== "function") {
@@ -114,7 +106,16 @@ export function makeClassComponentObserver(
         return this.render()
     }
 
-    patch(prototype, "componentDidMount", function () {
+    const originalComponentDidMount = prototype.componentDidMount
+    prototype.componentDidMount = function () {
+        if (__DEV__ && this.componentDidMount !== Object.getPrototypeOf(this).componentDidMount) {
+            const displayName = getDisplayName(componentClass)
+            throw new Error(
+                `[mobx-react] \`observer(${displayName}).componentDidMount\` must be defined on prototype.` +
+                    `\n\`componentDidMount = () => {}\` or \`componentDidMount = function() {}\` is not supported.`
+            )
+        }
+
         // `componentDidMount` may not be called at all. React can abandon the instance after `render`.
         // That's why we use finalization registry to dispose reaction created during render.
         // Happens with `<Suspend>` see #3492
@@ -140,16 +141,23 @@ export function makeClassComponentObserver(
         // As an alternative we could have `admin.instanceRef = new WeakRef(this)`, but lets avoid it if possible.
         admin.forceUpdate = () => this.forceUpdate()
 
-        if (!admin.reaction) {
+        if (!admin.reaction || admin.reactionInvalidatedBeforeMount) {
+            // Missing reaction:
             // 1. Instance was unmounted (reaction disposed) and immediately remounted without running render #3395.
             // 2. Reaction was disposed by finalization registry before mount. Shouldn't ever happen for class components:
             // `componentDidMount` runs synchronously after render, but our registry are deferred (can't run in between).
             // In any case we lost subscriptions to observables, so we have to create new reaction and re-render to resubscribe.
             // The reaction will be created lazily by following render.
+
+            // Reaction invalidated before mount:
+            // 1. A descendant's `componenDidMount` invalidated it's parent #3730
+
             admin.forceUpdate()
         }
-    })
+        return originalComponentDidMount?.apply(this, arguments)
+    }
 
+    // TODO@major Overly complicated "patch" is only needed to support the deprecated @disposeOnUnmount
     patch(prototype, "componentWillUnmount", function () {
         if (isUsingStaticRendering()) {
             return
@@ -159,6 +167,7 @@ export function makeClassComponentObserver(
         admin.reaction = null
         admin.forceUpdate = null
         admin.mounted = false
+        admin.reactionInvalidatedBeforeMount = false
     })
 
     return componentClass
@@ -207,28 +216,20 @@ function createReactiveRender(originalRender: any) {
 
 function createReaction(admin: ObserverAdministration) {
     return new Reaction(`${admin.name}.render()`, () => {
-        if (admin.isUpdating) {
-            // Reaction is suppressed when setting new state/props/context,
-            // this is when component is already being updated.
-            return
-        }
-
         if (!admin.mounted) {
             // This is neccessary to avoid react warning about calling forceUpdate on component that isn't mounted yet.
             // This happens when component is abandoned after render - our reaction is already created and reacts to changes.
-            // Due to the synchronous nature of `componenDidMount`, we don't have to worry that component could eventually mount and require update.
+            // `componenDidMount` runs synchronously after `render`, so unlike functional component, there is no delay during which the reaction could be invalidated.
+            // However `componentDidMount` runs AFTER it's descendants' `componentDidMount`, which CAN invalidate the reaction, see #3730. Therefore remember and forceUpdate on mount.
+            admin.reactionInvalidatedBeforeMount = true
             return
         }
 
         try {
-            // forceUpdate sets new `props`, since we made it observable, it would `reportChanged`, causing a loop.
-            admin.isUpdating = true
             admin.forceUpdate?.()
         } catch (error) {
             admin.reaction?.dispose()
             admin.reaction = null
-        } finally {
-            admin.isUpdating = false
         }
     })
 }
@@ -251,40 +252,23 @@ function observerSCU(nextProps: ClassAttributes<any>, nextState: any): boolean {
 }
 
 function createObservablePropDescriptor(key: "props" | "state" | "context") {
-    const atomKey = `${key}Atom`
     return {
         configurable: true,
         enumerable: true,
         get() {
             const admin = getAdministration(this)
-
-            let prevReadState = _allowStateReadsStart(true)
-
-            admin[atomKey].reportObserved()
-
-            _allowStateReadsEnd(prevReadState)
-
+            const derivation = _getGlobalState().trackingDerivation
+            if (derivation && derivation !== admin.reaction) {
+                throw new Error(
+                    `[mobx-react] Cannot read "${admin.name}.${key}" in a reactive context, as it isn't observable.
+                    Please use component lifecycle method to copy the value into a local observable first.
+                    See https://github.com/mobxjs/mobx/blob/main/packages/mobx-react/README.md#note-on-using-props-and-state-in-derivations`
+                )
+            }
             return admin[key]
         },
         set(value) {
-            const admin = getAdministration(this)
-            // forceUpdate issued by reaction sets new props.
-            // It sets isUpdating to true to prevent loop.
-            if (!admin.isUpdating && !shallowEqual(admin[key], value)) {
-                admin[key] = value
-                // This notifies all observers including our component,
-                // but we don't want to cause `forceUpdate`, because component is already updating,
-                // therefore supress component reaction.
-                admin.isUpdating = true
-                admin[atomKey].reportChanged()
-                admin.isUpdating = false
-            } else {
-                admin[key] = value
-            }
+            getAdministration(this)[key] = value
         }
     }
 }
-
-const observablePropsDescriptor = createObservablePropDescriptor("props")
-const observableStateDescriptor = createObservablePropDescriptor("state")
-const observableContextDescriptor = createObservablePropDescriptor("context")